requirements.txt
fastapi==0.110.0
python-multipart==0.0.9
uvicorn==0.27.1
openai-whisper==20231117
torch==2.2.1

